Move geometri to own model-file and a utils-folder for geometry transformation
