+++ src/main/kotlin/com/example/ruteplanlegger/model/Geometri.kt
@@ -1,0 +1,11 @@
+package com.example.ruteplanlegger.model
+
+data class Geometri(
+    val type: String = "Point",
+    val coordinates: LatLong
+)
+
+data class LatLong(
+    val long: Double,
+    var lat: Double
+)
