+++ src/main/kotlin/com/example/ruteplanlegger/model/Rasteplass.kt
@@ -1,15 +1,5 @@
 package com.example.ruteplanlegger.model
 
-data class Geometri(
-    val type: String = "Point",
-    val coordinates: LatLong
-)
-
-data class LatLong(
-    val long: Double,
-    var lat: Double
-)
-
 data class Rasteplass(
     val id: Int = 0,
     val navn: String = "",
