+++ src/main/kotlin/com/example/ruteplanlegger/service/RasteplasserService.kt
@@ -1,27 +1,10 @@
 package com.example.ruteplanlegger.service
 
-import com.example.ruteplanlegger.model.Geometri
-import com.example.ruteplanlegger.model.LatLong
 import com.example.ruteplanlegger.model.Rasteplass
+import com.example.ruteplanlegger.service.utils.createGeometri
 import com.fasterxml.jackson.databind.JsonNode
-import org.locationtech.jts.geom.Geometry
-import org.locationtech.jts.io.WKTReader
 import org.springframework.stereotype.Service
 import org.springframework.web.reactive.function.client.WebClient
-
-fun createGeometri(geometri: JsonNode): Geometri {
-    val wktNode = geometri["wkt"]
-    val wkt = wktNode.asText()
-    val wktReader = WKTReader()
-    val geometry: Geometry = wktReader.read(wkt)
-    val centroid =
-        geometry.centroid //gir et Point som er ca. i midten av en Linestring/Polygon
-    val lat: Double = centroid.coordinate.y
-    val long: Double = centroid.coordinate.x
-    val latLongObject = LatLong(long, lat)
-
-    return Geometri(type = "Point", coordinates = latLongObject)
-}
 
 fun createMinimalResteplassObject(data: JsonNode): List<Rasteplass> {
     return data["objekter"].map { rasteplass ->
